Expose utils as part of SDK
- SDK will expose the util functions so users can easily build solana
  transacitons
- Add barrels for cleaner exports and imports
- src/contstants will use explicit file path to avoid circular
  dependency
