+++ src/model/orca/pool/orca-pool.ts
@@ -1,26 +1,32 @@
-import { Connection, Keypair, PublicKey, Signer, TransactionSignature } from "@solana/web3.js";
-import { OrcaPool, OrcaU64, Quote } from "../../..";
-import { defaultSlippagePercentage } from "../../../constants/orca-defaults";
-import { deserializeAccount } from "../../../utils/web3/deserialize-account";
-import { PoolTokenCount, getTokenCount } from "../../../utils/web3/get-token-count";
-import { OrcaPoolParams } from "./pool-types";
-import { QuotePoolParams, QuoteBuilderFactory } from "../../quote/quote-builder";
-import { PercentageUtils } from "../../utils/percentage";
-import TransactionBuilder from "../../../utils/web3/transactions/transaction-builder";
-import {
-  createSwapInstruction,
-  createApprovalInstruction,
-} from "../../../utils/web3/instructions/pool-instructions";
 import { u64 } from "@solana/spl-token";
-import { sendAndConfirmTransaction } from "../../../utils/web3/transactions/transactions";
-import { getTokens } from "../../../utils/pool-utils";
-import Decimal from "decimal.js";
-import { U64Utils } from "../../../utils/numbers/u64-utils";
-import { OrcaToken } from "../../../public/types";
 import {
+  Connection,
+  PublicKey,
+  Keypair,
+  TransactionSignature,
+  sendAndConfirmTransaction,
+} from "@solana/web3.js";
+import Decimal from "decimal.js";
+import { OrcaPool, OrcaToken, Quote } from "../../..";
+import { defaultSlippagePercentage } from "../../../constants/orca-defaults";
+import { PercentageUtils } from "../../../public/utils/percentage-utils";
+import {
+  OrcaU64,
   deriveAssociatedTokenAddress,
+  deserializeAccount,
+  getTokens,
+  U64Utils,
+  PoolTokenCount,
+  getTokenCount,
   resolveAssociatedTokenAddress,
-} from "../../../utils/web3/ata-utils";
+} from "../../../public";
+import {
+  createApprovalInstruction,
+  createSwapInstruction,
+} from "../../../public/utils/web3/instructions/pool-instructions";
+import TransactionBuilder from "../../../public/utils/web3/transactions/transaction-builder";
+import { QuotePoolParams, QuoteBuilderFactory } from "../../quote/quote-builder";
+import { OrcaPoolParams } from "./pool-types";
 
 export class OrcaPoolImpl implements OrcaPool {
   private connection: Connection;
