+++ src/public/main/types.ts
@@ -1,0 +1,9 @@
+import { OrcaPool, OrcaPoolConfig } from "..";
+
+export type Orca = {
+  /**
+   * Get an instance of an Orca pool.
+   * @param pool a pool config targeting an Orca pool
+   */
+  getPool: (pool: OrcaPoolConfig) => OrcaPool;
+};
