+++ src/public/pools/config.ts
