+++ src/public/utils/models/percentage.ts
@@ -1,0 +1,6 @@
+import { u64 } from "@solana/spl-token";
+
+export type Percentage = {
+  numerator: u64;
+  denominator: u64;
+};
