+++ src/public/utils/web3/instructions/pool-instructions.ts
@@ -1,9 +1,9 @@
 import { Token, TOKEN_PROGRAM_ID, u64 } from "@solana/spl-token";
 import { TokenSwap } from "@solana/spl-token-swap";
 import { Keypair, PublicKey } from "@solana/web3.js";
-import { OrcaPoolParams, OrcaPoolToken } from "../../../model/orca/pool/pool-types";
-import { Instruction } from "../../../model/utils/instruction";
+import { OrcaPoolParams, OrcaPoolToken } from "../../../../model/orca/pool/pool-types";
 import { ORCA_TOKEN_SWAP_ID } from "../../constants";
+import { Instruction } from "../../models";
 
 export const createApprovalInstruction = (
   ownerAddress: PublicKey,
